backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-chess==1.999
stockfish==3.28.0
anthropic==0.42.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
supabase==2.10.0
